Fix CI/CD: Add size_t cast in Labellist::ensureCapacity
Fix Clang-15 compilation error with -Wsign-conversion:
- Cast int parameter to size_t when calling vec.reserve()
- Clang is stricter than GCC about implicit sign conversions
- Consistent with other Labellist methods (reserve, operator[])

+++ src/compiler/lparser.h
@@ -237,7 +237,7 @@ public:
   /* For luaM_growvector replacement */
   inline void ensureCapacity(int needed) {
     if (needed > getSize()) {
-      vec.reserve(needed);
+      vec.reserve(static_cast<size_t>(needed));
     }
   }
   inline Labeldesc* allocateNew() {
